Final Commit

+++ src/components/WorkoutDetails.js
@@ -1,15 +1,27 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
+  const {user} = useAuthContext()
   const { dispatch } = useWorkoutsContext()
 
   const handleClick = async () => {
+
+    if(!user)
+    {
+      return
+    }
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
+    console.log("Delet API")
+
     const json = await response.json()
 
     if (response.ok) {
